// LinkDev.DataMigration.FrontEnd/src/app/model/configuration.model.ts
export class ValueMapping implements IValueMapping
{
	isIgnoreValues?: boolean | undefined;
	sourceField: string;
	sourceFieldType: number;
	sourceValue?: string | undefined;
	destinationField: string;
	destinationFieldType?: number | undefined;
	destinationValue?: string | undefined;

	constructor(data?: IValueMapping)
	{
		if (data)
		{
			for (var property in data)
			{
				if (data.hasOwnProperty(property))
					(<any>this)[property] = (<any>data)[property];
			}
		}
	}

	init(data?: any)
	{
		if (data)
		{
			this.isIgnoreValues = data["isIgnoreValues"];
			this.sourceField = data["sourceField"];
			this.sourceFieldType = data["sourceFieldType"];
			this.sourceValue = data["sourceValue"];
			this.destinationField = data["destinationField"];
			this.destinationFieldType = data["destinationFieldType"];
			this.destinationValue = data["destinationValue"];
		}
	}

	static fromJS(data: any): ValueMapping
	{
		let result = new ValueMapping();
		result.init(data);
		return result;
	}

	toJSON(data?: any)
	{
		data = typeof data === 'object' ? data : {};
		data["isIgnoreValues"] = this.isIgnoreValues;
		data["sourceField"] = this.sourceField;
		data["sourceFieldType"] = this.sourceFieldType;
		data["sourceValue"] = this.sourceValue;
		data["destinationField"] = this.destinationField;
		data["destinationFieldType"] = this.destinationFieldType;
		data["destinationValue"] = this.destinationValue;
		return data;
	}
}

export interface IValueMapping
{
	isIgnoreValues?: boolean | undefined;
	sourceField: string;
	sourceFieldType: number;
	sourceValue?: string | undefined;
	destinationField: string;
	destinationFieldType?: number | undefined;
	destinationValue?: string | undefined;
}
// ...
export class ValuesMapping implements IValuesMapping
{
	targetEntityName?: string | undefined;
	valueMappings: ValueMapping[] = [];

	constructor(data?: IValuesMapping)
	{
		if (data)
		{
			for (var property in data)
			{
				if (data.hasOwnProperty(property))
					(<any>this)[property] = (<any>data)[property];
			}
		}
	}

	init(data?: any)
	{
		if (data)
		{
			this.targetEntityName = data["targetEntityName"];
			if (data["valueMappings"] && data["valueMappings"].constructor === Array)
			{
				this.valueMappings = [];
				for (let item of data["valueMappings"])
					this.valueMappings.push(ValueMapping.fromJS(item));
			}
		}
	}

	static fromJS(data: any): ValuesMapping
	{
		let result = new ValuesMapping();
		result.init(data);
		return result;
	}

	toJSON(data?: any)
	{
		data = typeof data === 'object' ? data : {};
		data["targetEntityName"] = this.targetEntityName;
		if (this.valueMappings && this.valueMappings.constructor === Array)
		{
			data["valueMappings"] = [];
			for (let item of this.valueMappings)
				data["valueMappings"].push(item.toJSON());
		}
		return data;
	}
}

export interface IValuesMapping
{
	targetEntityName?: string | undefined;
	valueMappings: ValueMapping[];
}